Issue edina/spatial-memories/issues/60: Added new theme layout - for themes water, fieldtripgb and default.
Note that the images folder in the water theme is symlinked to from everywhere else so if you require new images
this should be used.

Also note that the theme name in settings.html should correspond to the theme path in the theme directory to ensure it
is loaded correctly.